Moved the search matching from common/util to main/search
