+++ src/main/modules/folders.js
@@ -1,5 +1,5 @@
 
-import { generateSearchRegex, stringMatchQuality } from "../../common/util";
+import { generateSearchRegex, stringMatchQuality } from "../search";
 import { config } from "../config";
 import { stat, exists, readdir } from "fs";
 import path, { extname } from 'path';
