+++ src/main/modules/linux-windows.js
@@ -1,5 +1,5 @@
 
-import { stringMatchQuality } from '../../common/util';
+import { stringMatchQuality } from '../search';
 import { exec } from 'child_process';
 
 const LinuxWindowsModule = {
